fastapi/examples/export-data/requirements.txt
fastapi
uvicorn
pandas
xlsxwriter
reportlab
pyarrow
mysql-connector-python
